feat: 앱 버전 확인 분기 추가
