+++ android/app/src/main/java/com/mulkkam/ui/main/MainViewModel.kt
@@ -7,6 +7,7 @@ import com.mulkkam.di.RepositoryInjection.devicesRepository
 import com.mulkkam.di.RepositoryInjection.healthRepository
 import com.mulkkam.di.RepositoryInjection.membersRepository
 import com.mulkkam.di.RepositoryInjection.tokenRepository
+import com.mulkkam.di.RepositoryInjection.versionsRepository
 import com.mulkkam.di.WorkInjection.calorieScheduler
 import com.mulkkam.domain.work.CalorieScheduler.Companion.DEFAULT_CHECK_CALORIE_INTERVAL_HOURS
 import com.mulkkam.ui.util.MutableSingleLiveData
@@ -25,6 +26,9 @@ class MainViewModel : ViewModel() {
     private val _onFirstLaunch: MutableSingleLiveData<Unit> = MutableSingleLiveData()
     val onFirstLaunch: SingleLiveData<Unit>
         get() = _onFirstLaunch
+
+    private val _isAppOutdated: MutableSingleLiveData<Boolean> = MutableSingleLiveData()
+    val isAppOutdated: SingleLiveData<Boolean> get() = _isAppOutdated
 
     init {
         getFcmToken()
@@ -76,5 +80,37 @@ class MainViewModel : ViewModel() {
                 tokenRepository.deleteFcmToken()
             }
         }
+    }
+
+    fun checkAppVersion(currentVersionName: String) {
+        viewModelScope.launch {
+            runCatching {
+                versionsRepository.getMinimumVersion().getOrError()
+            }.onSuccess { minimumVersion ->
+                _isAppOutdated.setValue(isOutdated(currentVersionName, minimumVersion))
+            }.onFailure {
+                _isAppOutdated.setValue(true)
+            }
+        }
+    }
+
+    private fun isOutdated(
+        currentVersion: String,
+        minimumVersion: String,
+    ): Boolean {
+        val currentParts: List<Int> = currentVersion.split(".").mapNotNull { it.toIntOrNull() }
+        val minimumParts: List<Int> = minimumVersion.split(".").mapNotNull { it.toIntOrNull() }
+        val maxLength: Int = maxOf(currentParts.size, minimumParts.size)
+
+        for (index in 0 until maxLength) {
+            val currentPart: Int = currentParts.getOrElse(index) { 0 }
+            val minimumPart: Int = minimumParts.getOrElse(index) { 0 }
+
+            when {
+                currentPart < minimumPart -> return true
+                currentPart > minimumPart -> return false
+            }
+        }
+        return false
     }
 }
