+++ android/app/src/main/java/com/mulkkam/ui/util/extensions/ContextExtensions.kt
@@ -2,10 +2,14 @@ package com.mulkkam.ui.util.extensions
 
 import android.content.Context
 import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
 import androidx.annotation.StringRes
 import androidx.core.net.toUri
 import androidx.health.connect.client.HealthConnectClient
 import com.mulkkam.di.HealthConnectInjection.PROVIDER_PACKAGE_NAME
+
+private const val MIN_VERSION_NAME: String = "0.0.0"
 
 /**
  * Health Connect SDK 설치/업데이트 상태를 확인한다.
@@ -43,3 +47,18 @@ fun Context.openTermsLink(
         )
     startActivity(intent)
 }
+
+fun Context.getAppVersion(): String =
+    runCatching {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            val packageInfo =
+                packageManager.getPackageInfo(
+                    packageName,
+                    PackageManager.PackageInfoFlags.of(0),
+                )
+            packageInfo.versionName
+        } else {
+            val packageInfo = packageManager.getPackageInfo(packageName, 0)
+            packageInfo.versionName
+        }
+    }.getOrNull() ?: MIN_VERSION_NAME
